requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
mysql-connector-python==8.0.33
python-dotenv
random-user-agent